labeltosvg/requirement.txt
more-itertools==10.7.0
svg.py==1.6.0
